Changed helper to use size_t rather than int

+++ 0x1C-binary_trees/15-binary_tree_is_full.c
@@ -7,7 +7,7 @@
 * Return: 1 if full, 0 otherwise
 */
 
-int binary_tree_is_full_help(const binary_tree_t *tree);
+size_t binary_tree_is_full_help(const binary_tree_t *tree);
 
 int binary_tree_is_full(const binary_tree_t *tree)
 {
@@ -27,7 +27,7 @@ int binary_tree_is_full(const binary_tree_t *tree)
 * Return: 0 if full, amount of non full nodes otherwise
 */
 
-int binary_tree_is_full_help(const binary_tree_t *tree)
+size_t binary_tree_is_full_help(const binary_tree_t *tree)
 {
 	if (tree->left != NULL && tree->right != NULL)
 		return ((0 + binary_tree_is_full_help(tree->left)) +
